requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==36.0.1
Deprecated==1.2.13
Django==3.2.5
django-cors-headers==3.11.0
django-oauth-toolkit==1.7.0
djangorestframework==3.12.4
dnspython==2.1.0
idna==3.3
jwcrypto==1.0
oauthlib==3.2.0
Pillow==8.3.2
psycopg2==2.9
pycparser==2.21
pymongo==3.12.0
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
urllib3==1.26.7
wrapt==1.13.3
